// types.d.ts
export interface Coordinates {
	/**
	 * An x-coordinate on the board (integer).
	 * Example: 1
	 */
	x: number;

	/**
	 * An y-coordinate on the board (integer).
	 * Example: 1
	 */
	y: number;
}
// ...
export type Direction = 'up' | 'left' | 'down' | 'right';